+++ fiboa_cli/datasets/ch.py
@@ -1,0 +1,73 @@
+from ..convert_utils import convert as convert_
+import pandas as pd
+import numpy as np
+
+SOURCES = None
+DATA_ACCESS = """
+Data must be obtained from the Swiss open data portal at https://www.geodienste.ch/services/lwb_nutzungsflaechen .
+
+One can filter on "Verfügbarkeit" == "Frei erhältlich" to select only the open data.
+That leaves out Cantons AR, NW, OW, VD and LI as on this date (2014-11-12).
+The downloaded data can be shared with a open_by license. See https://opendata.swiss/de/terms-of-use .
+
+Use the `-i` CLI parameter to provide the data source.
+Download the Open data response to a local gpkg file (use `.gpkg` as file extension).
+
+fiboa convert ch -o swiss.parquet -i lwb_nutzungsflaechen_lv95/geopackage/lwb_nutzungsflaechen_v2_0_lv95.gpkg
+"""
+
+ID = "ch"
+SHORT_NAME = "Switzerland"
+TITLE = "Field boundaries for Switzerland"
+DESCRIPTION = """The cropfields of Switzerland (Nutzungsflächen) are published per administrative subdivision called Canton."""
+PROVIDERS = [
+    {
+        "name": "Konferenz der kantonalen Geoinformations- und Katasterstellen",
+        "url": "https://www.kgk-cgc.ch/",
+        "roles": ["producer", "licensor"]
+    }
+]
+LICENSE = {
+    "title": "opendata.swiss terms of use",
+    "href": "https://opendata.swiss/en/terms-of-use",
+    "type": "text/html",
+    "rel": "license"
+}
+COLUMNS = {
+    'geometry': 'geometry',
+    'id': 'id',
+    'flaeche_m2': 'area',
+    'kanton': 'administrative_area_level_1',
+    "nutzung": "crop_name",
+    'bezugsjahr': 'determination_datetime'
+}
+
+COLUMN_MIGRATIONS = {
+    'flaeche_m2': lambda column: np.where(column>0, column/10000, 0.001),
+    'bezugsjahr': lambda col: pd.to_datetime(col, format='%Y')
+}
+
+MISSING_SCHEMAS = {
+    'properties': {
+        'administrative_area_level_1': {
+            'type': 'string'
+        },
+        'crop_name': {
+            'type': 'string'
+        },
+    }
+}
+
+def convert(output_file, cache = None, **kwargs):
+    convert_(
+        output_file, cache, SOURCES,
+        COLUMNS, ID, TITLE, DESCRIPTION,
+        license=LICENSE,
+        missing_schemas=MISSING_SCHEMAS,
+        column_migrations=COLUMN_MIGRATIONS,
+        providers=PROVIDERS,
+        explode_multipolygon=True,
+        index_as_id=True,
+        fid_as_index=True,
+        **kwargs
+    )